oss-config/pyproject.toml
[project]
name = "meeting-oss-uploader"
version = "0.1.0"
description = "Upload meeting videos to OSS and transcribe with Qwen ASR"
requires-python = ">=3.8"
dependencies = [
    "oss2>=2.18.0",
    "python-dotenv>=1.0.0",
    "dashscope>=1.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["*.py"]
